ログイン機能実装

+++ app/Http/Controllers/Tweet/DeleteController.php
@@ -4,16 +4,21 @@ namespace App\Http\Controllers\Tweet;
 
 use App\Http\Controllers\Controller;
 use App\Models\Tweet;
+use App\Service\TweetService;
 use Illuminate\Http\Request;
+use Symfony\Component\Finder\Exception\AccessDeniedException;
 
 class DeleteController extends Controller
 {
     /**
      * Handle the incoming request.
      */
-    public function __invoke(Request $request)
+    public function __invoke(Request $request, TweetService $tweetService)
     {
         $tweetId = (int) $request->route('tweetId');
+        if (!$tweetService->checkOwnTweet($request->user()->id, $tweetId)) {
+            throw new AccessDeniedException;
+        }
         $tweet = Tweet::where('id', $tweetId)->firstOrFail();
         $tweet->delete();
         return redirect()->route('tweet.index')->with('feedback.success', "つぶやきを削除しました");
